# schemas/generated/orchestration-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tasker Web API",
    "description": "REST API for the Tasker workflow orchestration system",
    "contact": {
      "name": "Tasker Systems",
      "email": "support@tasker-systems.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "API version 1"
    }
  ],
  "paths": {
...
    "/v1/tasks/{uuid}/workflow_steps": {
      "get": {
        "tags": [
          "workflow_steps"
        ],
        "summary": "List workflow steps for a task: GET /v1/tasks/{uuid}/workflow_steps",
        "description": "**Required Permission:** `steps:read`",
        "operationId": "list_task_steps",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "description": "Task UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of workflow steps",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StepResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid task UUID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "api_key_auth": []
          }
        ],
        "x-required-permission": "steps:read"
      }
    },
    "/v1/tasks/{uuid}/workflow_steps/{step_uuid}": {
      "get": {
        "tags": [
          "workflow_steps"
        ],
        "summary": "Get workflow step details: GET /v1/tasks/{uuid}/workflow_steps/{step_uuid}",
        "description": "**Required Permission:** `steps:read`",
        "operationId": "get_step",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "description": "Task UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "step_uuid",
            "in": "path",
            "description": "Step UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workflow step details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StepResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid UUID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Step not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "api_key_auth": []
          }
        ],
        "x-required-permission": "steps:read"
      },
      "patch": {
        "tags": [
          "workflow_steps"
        ],
        "summary": "Manually resolve or complete a workflow step: PATCH /v1/tasks/{uuid}/workflow_steps/{step_uuid}",
        "description": "This endpoint supports two resolution modes:\n\n1. **Simple Resolution** (`ResolveManually`): Marks the step as `resolved_manually` without providing execution results.\n   Use this when you want to acknowledge a step issue without continuing workflow execution.\n\n2. **Complete with Results** (`CompleteManually`): Provides execution results and transitions the step to `complete` state.\n   - Operators provide only `result` (business data) and optional `metadata`\n   - System automatically enforces `success: true` and `status: \"completed\"`\n   - Prevents operators from incorrectly setting `status` or `error` fields\n   - Dependent steps can consume the provided results, allowing workflow continuation\n\nYou can optionally reset the attempt counter when resolving a step that has exceeded retry limits.\n\n**Required Permission:** `steps:resolve`",
        "operationId": "resolve_step_manually",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "description": "Task UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "step_uuid",
            "in": "path",
            "description": "Step UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StepManualAction"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Step resolved successfully. The response indicates whether the step was marked as 'resolved_manually' (simple resolution) or 'complete' (with execution results).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StepResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or step cannot be resolved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Step not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "api_key_auth": []
          }
        ],
        "x-required-permission": "steps:resolve"
      }
    },
    "/v1/tasks/{uuid}/workflow_steps/{step_uuid}/audit": {
      "get": {
        "tags": [
          "workflow_steps"
        ],
        "summary": "Get audit history for a workflow step: GET /v1/tasks/{uuid}/workflow_steps/{step_uuid}/audit",
        "description": "Returns SOC2-compliant audit trail for step execution results (TAS-62).\nEach audit record includes:\n- Worker attribution (worker_uuid, correlation_id)\n- Execution summary (success, execution_time_ms)\n- Full execution result via JOIN to transition metadata\n\n## Response\n\nReturns an array of audit records ordered by recorded_at DESC (most recent first).\nEach record links to the transition that captured the execution result.\n\n**Required Permission:** `steps:read`",
        "operationId": "get_step_audit",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "description": "Task UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "step_uuid",
            "in": "path",
            "description": "Step UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Audit history for the step",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StepAuditResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid UUID format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Step not found or does not belong to task",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "api_key_auth": []
          }
        ],
        "x-required-permission": "steps:read"
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "NotFound"
            ],
            "properties": {
              "NotFound": {
                "type": "object",
                "required": [
                  "message"
                ],
                "properties": {
                  "message": {
                    "type": "string"
                  }
                }
              }
            }
          },
          {
            "type": "string",
            "enum": [
              "Forbidden"
            ]
          },
          {
            "type": "string",
            "enum": [
              "Unauthorized"
            ]
          },
          {
            "type": "object",
            "required": [
              "BadRequest"
            ],
            "properties": {
              "BadRequest": {
                "type": "object",
                "required": [
                  "message"
                ],
                "properties": {
                  "message": {
                    "type": "string"
                  }
                }
              }
            }
          },
          {
            "type": "object",
            "description": "TAS-154: Conflict - resource already exists (e.g., duplicate identity hash)",
            "required": [
              "Conflict"
            ],
            "properties": {
              "Conflict": {
                "type": "object",
                "description": "TAS-154: Conflict - resource already exists (e.g., duplicate identity hash)",
                "required": [
                  "message"
                ],
                "properties": {
                  "message": {
                    "type": "string"
                  }
                }
              }
            }
          },
          {
            "type": "string",
            "enum": [
              "ServiceUnavailable"
            ]
          },
          {
            "type": "string",
            "enum": [
              "Timeout"
            ]
          },
          {
            "type": "string",
            "enum": [
              "CircuitBreakerOpen"
            ]
          },
          {
            "type": "object",
            "description": "TAS-75: System backpressure active - includes Retry-After header",
            "required": [
              "Backpressure"
            ],
            "properties": {
              "Backpressure": {
                "type": "object",
                "description": "TAS-75: System backpressure active - includes Retry-After header",
                "required": [
                  "reason",
                  "retry_after_seconds"
                ],
                "properties": {
                  "reason": {
                    "type": "string"
                  },
                  "retry_after_seconds": {
                    "type": "integer",
                    "format": "int64",
                    "minimum": 0
                  }
                }
              }
            }
          },
          {
            "type": "object",
            "required": [
              "DatabaseError"
            ],
            "properties": {
              "DatabaseError": {
                "type": "object",
                "required": [
                  "operation"
                ],
                "properties": {
                  "operation": {
                    "type": "string"
                  }
                }
              }
            }
          },
          {
            "type": "object",
            "required": [
              "AuthenticationError"
            ],
            "properties": {
              "AuthenticationError": {
                "type": "object",
                "required": [
                  "reason"
                ],
                "properties": {
                  "failure_context": {
                    "oneOf": [
                      {
                        "type": "null"
                      },
                      {
                        "$ref": "#/components/schemas/AuthFailureContext",
                        "description": "Machine-readable failure context for gateway signaling headers"
                      }
                    ]
                  },
                  "reason": {
                    "type": "string"
                  }
                }
              }
            }
          },
          {
            "type": "object",
            "required": [
              "AuthorizationError"
            ],
            "properties": {
              "AuthorizationError": {
                "type": "object",
                "required": [
                  "reason"
                ],
                "properties": {
                  "failure_context": {
                    "oneOf": [
                      {
                        "type": "null"
                      },
                      {
                        "$ref": "#/components/schemas/AuthFailureContext",
                        "description": "Machine-readable failure context for gateway signaling headers"
                      }
                    ]
                  },
                  "reason": {
                    "type": "string"
                  }
                }
              }
            }
          },
          {
            "type": "object",
            "required": [
              "InvalidUuid"
            ],
            "properties": {
              "InvalidUuid": {
                "type": "object",
                "required": [
                  "uuid"
                ],
                "properties": {
                  "uuid": {
                    "type": "string"
                  }
                }
              }
            }
          },
          {
            "type": "string",
            "enum": [
              "JsonError"
            ]
          },
          {
            "type": "string",
            "enum": [
              "Internal"
            ]
          }
        ],
        "description": "Web API specific errors with HTTP status code mappings"
      },
      "AuthFailureContext": {
        "type": "object",
        "description": "Auth failure context for gateway signaling.\n\nIncluded in 401/403 response headers so upstream gateways (ALB, NLB, API Gateway)\ncan make informed rate-limiting decisions based on failure severity.",
        "required": [
          "reason",
          "severity"
        ],
        "properties": {
          "reason": {
            "$ref": "#/components/schemas/AuthFailureReason",
            "description": "Machine-readable failure category"
          },
          "severity": {
            "$ref": "#/components/schemas/AuthFailureSeverity",
            "description": "Suggested severity for gateway action"
          }
        }
      },
      "AuthFailureReason": {
        "type": "string",
        "description": "Machine-readable auth failure reasons.",
        "enum": [
          "Missing",
          "Expired",
          "Invalid",
          "Malformed",
          "Forbidden"
        ]
      },
...
      "StepAuditResponse": {
        "type": "object",
        "description": "Step audit record response for SOC2-compliant audit trails (TAS-62)\n\nProvides attribution context (worker_uuid, correlation_id) and execution summary\nwith full results retrieved via JOIN to the transitions table.\n\n## Usage\n\n```json\nGET /v1/tasks/{uuid}/workflow_steps/{step_uuid}/audit\n[\n  {\n    \"audit_uuid\": \"01934567-89ab-cdef-0123-456789abcdef\",\n    \"workflow_step_uuid\": \"...\",\n    \"transition_uuid\": \"...\",\n    \"task_uuid\": \"...\",\n    \"recorded_at\": \"2025-12-03T10:30:00Z\",\n    \"worker_uuid\": \"worker-abc-123\",\n    \"correlation_id\": \"corr-xyz-789\",\n    \"success\": true,\n    \"execution_time_ms\": 150,\n    \"result\": { ... },\n    \"step_name\": \"validate_order\",\n    \"from_state\": \"in_progress\",\n    \"to_state\": \"enqueued_for_orchestration\"\n  }\n]\n```",
        "required": [
          "audit_uuid",
          "workflow_step_uuid",
          "transition_uuid",
          "task_uuid",
          "recorded_at",
          "success",
          "step_name",
          "to_state"
        ],
        "properties": {
          "audit_uuid": {
            "type": "string",
            "description": "Primary key of the audit record"
          },
          "correlation_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Correlation ID for distributed tracing"
          },
          "execution_time_ms": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Execution time in milliseconds"
          },
          "from_state": {
            "type": [
              "string",
              "null"
            ],
            "description": "State before the transition"
          },
          "recorded_at": {
            "type": "string",
            "description": "When the audit record was created (ISO 8601 format)"
          },
          "result": {
            "description": "Full execution result from the transition"
          },
          "step_name": {
            "type": "string",
            "description": "Name of the step"
          },
          "success": {
            "type": "boolean",
            "description": "Whether the step execution succeeded"
          },
          "task_uuid": {
            "type": "string",
            "description": "The parent task"
          },
          "to_state": {
            "type": "string",
            "description": "State after the transition"
          },
          "transition_uuid": {
            "type": "string",
            "description": "The specific transition that recorded the result"
          },
          "worker_uuid": {
            "type": [
              "string",
              "null"
            ],
            "description": "UUID of the worker instance that processed this step"
          },
          "workflow_step_uuid": {
            "type": "string",
            "description": "The workflow step this audit record belongs to"
          }
        }
      },
...
      "StepResponse": {
        "type": "object",
        "description": "Step details response with readiness information",
        "required": [
          "step_uuid",
          "task_uuid",
          "name",
          "created_at",
          "updated_at",
          "current_state",
          "dependencies_satisfied",
          "retry_eligible",
          "ready_for_execution",
          "total_parents",
          "completed_parents",
          "attempts",
          "max_attempts"
        ],
        "properties": {
          "attempts": {
            "type": "integer",
            "format": "int32"
          },
          "completed_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "completed_parents": {
            "type": "integer",
            "format": "int32"
          },
          "created_at": {
            "type": "string"
          },
          "current_state": {
            "type": "string"
          },
          "dependencies_satisfied": {
            "type": "boolean"
          },
          "last_attempted_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "last_failure_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "max_attempts": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "next_retry_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "ready_for_execution": {
            "type": "boolean"
          },
          "results": {},
          "retry_eligible": {
            "type": "boolean"
          },
          "step_uuid": {
            "type": "string"
          },
          "task_uuid": {
            "type": "string"
          },
          "total_parents": {
            "type": "integer",
            "format": "int32"
          },
          "updated_at": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key_auth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}